refactor: rename typeInternal to type.internal
